backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiofiles=0.8.0=pypi_0
anyio=3.5.0=pypi_0
asgiref=3.5.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
certifi=2021.10.8=py37haa95532_2
click=8.1.2=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
debugpy=1.5.1=py37hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
fastapi=0.75.2=pypi_0
h11=0.13.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
ipykernel=6.9.1=py37haa95532_0
ipython=7.31.1=py37haa95532_0
jedi=0.18.1=py37haa95532_1
jinja2=3.1.1=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.9.2=py37haa95532_0
markupsafe=2.1.1=pypi_0
matlabengineforpython=R2020a=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
nest-asyncio=1.5.5=py37haa95532_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.2.4=py37haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
pydantic=1.9.0=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
python=3.7.0=hea74fb7_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pywin32=302=py37h827c3e9_1
pyzmq=22.3.0=py37hd77b12b_2
setuptools=61.2.0=py37haa95532_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=pypi_0
spyder-kernels=2.1.3=py37haa95532_0
starlette=0.17.1=pypi_0
tornado=6.1=py37h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.2.0=pypi_0
uvicorn=0.17.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py37haa95532_2
zipp=3.8.0=pypi_0

pymysql==1.0.2
uvicorn==0.17.6
fastapi==0.75.2
starlette==0.17.1